cherry pick 01

+++ src/Rendering/Cameras/PhysicallyBasedCamera.h
@@ -10,7 +10,7 @@ public:
   PhysicallyBasedCamera(float _focalDistance, float _lensRadius, float _fov)
       : m_FocalDistance(_focalDistance), m_LensRadius(_lensRadius),
         PinholeCamera(_fov){};
-  virtual DirectX::SimpleMath::Ray GetRay(int _x, int _y, int _w, int _h,
+  virtual DirectX::SimpleMath::Ray GetRay(float _x, float _y, int _w, int _h,
                                           std::default_random_engine &_rnd,
                                           float &weight) const override;
 };
