+++ src/Rendering/Cameras/PinholeCamera.h
@@ -7,7 +7,7 @@ protected:
 
 public:
   PinholeCamera(float _fov) : m_FOV(_fov){};
-  virtual DirectX::SimpleMath::Ray GetRay(int _x, int _y, int _w, int _h,
+  virtual DirectX::SimpleMath::Ray GetRay(float _x, float _y, int _w, int _h,
                                           std::default_random_engine &_rnd,
                                           float &weight) const override;
 };
